First step towards #315
