+++ modules/src/ics04_channel/msgs/acknowledgement.rs
@@ -1,0 +1,72 @@
+use crate::ics04_channel::error::{Error, Kind};
+use crate::ics04_channel::packet::Packet;
+use crate::ics23_commitment::commitment::CommitmentProof;
+use crate::{proofs::Proofs, tx_msg::Msg, Height};
+
+use tendermint::account::Id as AccountId;
+
+use serde_derive::{Deserialize, Serialize};
+
+/// Message type for the `MsgAcknowledgement` message.
+const TYPE_MSG_ACKNOWLEDGEMENT: &str = "ics04/opaque";
+
+///
+/// Message definition for packet acknowledgements.
+///
+#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
+pub struct MsgAcknowledgement {
+    packet: Packet,
+    acknowledgement: Vec<u8>,
+    proofs: Proofs,
+    signer: AccountId,
+}
+
+impl MsgAcknowledgement {
+    pub fn new(
+        packet: Packet,
+        acknowledgement: Vec<u8>,
+        proof: CommitmentProof,
+        proof_height: Height,
+        signer: AccountId,
+    ) -> Result<MsgAcknowledgement, Error> {
+        if acknowledgement.len() > 100 {
+            return Err(Kind::AcknowledgementTooLong.into());
+        }
+
+        Ok(Self {
+            packet: packet
+                .validate()
+                .map_err(|e| Kind::InvalidPacket.context(e))?,
+            acknowledgement,
+            proofs: Proofs::new(proof, None, None, proof_height)
+                .map_err(|e| Kind::InvalidProof.context(e))?,
+            signer,
+        })
+    }
+}
+
+impl Msg for MsgAcknowledgement {
+    type ValidationError = Error;
+
+    fn route(&self) -> String {
+        crate::keys::ROUTER_KEY.to_string()
+    }
+
+    fn get_type(&self) -> String {
+        TYPE_MSG_ACKNOWLEDGEMENT.to_string()
+    }
+
+    fn validate_basic(&self) -> Result<(), Self::ValidationError> {
+        // Nothing to validate
+        // All the validation is performed on creation
+        Ok(())
+    }
+
+    fn get_sign_bytes(&self) -> Vec<u8> {
+        todo!()
+    }
+
+    fn get_signers(&self) -> Vec<AccountId> {
+        vec![self.signer]
+    }
+}
